test float -> double

+++ src/gd.c
@@ -3477,18 +3477,18 @@ BGD_DECLARE(void) gdImageCopyResampled (gdImagePtr dst,
 	}
 	for (y = dstY; (y < dstY + dstH); y++) {
 		for (x = dstX; (x < dstX + dstW); x++) {
-			float sy1, sy2, sx1, sx2;
-			float sx, sy;
-			float spixels = 0.0;
-			float red = 0.0, green = 0.0, blue = 0.0, alpha = 0.0;
-			float alpha_factor, alpha_sum = 0.0, contrib_sum = 0.0;
-			sy1 = ((float)(y - dstY)) * (float)srcH / (float)dstH;
-			sy2 = ((float)(y + 1 - dstY)) * (float) srcH / (float) dstH;
+			double sy1, sy2, sx1, sx2;
+			double sx, sy;
+			double spixels = 0;
+			double red = 0.0, green = 0.0, blue = 0.0, alpha = 0.0;
+			double alpha_factor, alpha_sum = 0.0, contrib_sum = 0.0;
+			sy1 = ((double)(y - dstY)) * (double)srcH / (double)dstH;
+			sy2 = ((double)(y + 1 - dstY)) * (double) srcH / (double) dstH;
 			sy = sy1;
 			do {
-				float yportion;
+				double yportion;
 				if (floorf(sy) == floorf(sy1)) {
-					yportion = 1.0 - (sy - floorf(sy));
+					yportion = 1.0f - (sy - floorf(sy));
 					if (yportion > sy2 - sy1) {
 						yportion = sy2 - sy1;
 					}
@@ -3496,17 +3496,17 @@ BGD_DECLARE(void) gdImageCopyResampled (gdImagePtr dst,
 				} else if (sy == floorf(sy2)) {
 					yportion = sy2 - floorf(sy2);
 				} else {
-					yportion = 1.0;
+					yportion = 1.0f;
 				}
-				sx1 = ((float)(x - dstX)) * (float) srcW / dstW;
-				sx2 = ((float)(x + 1 - dstX)) * (float) srcW / dstW;
+				sx1 = ((double)(x - dstX)) * (double) srcW / dstW;
+				sx2 = ((double)(x + 1 - dstX)) * (double) srcW / dstW;
 				sx = sx1;
 				do {
-					float xportion;
-					float pcontribution;
+					double xportion;
+					double pcontribution;
 					int p;
 					if (floorf(sx) == floorf(sx1)) {
-						xportion = 1.0 - (sx - floorf(sx));
+						xportion = 1.0f - (sx - floorf(sx));
 						if (xportion > sx2 - sx1) {
 							xportion = sx2 - sx1;
 						}
@@ -3514,7 +3514,7 @@ BGD_DECLARE(void) gdImageCopyResampled (gdImagePtr dst,
 					} else if (sx == floorf(sx2)) {
 						xportion = sx2 - floorf(sx2);
 					} else {
-						xportion = 1.0;
+						xportion = 1.0f;
 					}
 					pcontribution = xportion * yportion;
 					p = gdImageGetTrueColorPixel(src, (int) sx + srcX, (int) sy + srcY);
@@ -3527,7 +3527,7 @@ BGD_DECLARE(void) gdImageCopyResampled (gdImagePtr dst,
 					alpha_sum += alpha_factor;
 					contrib_sum += pcontribution;
 					spixels += xportion * yportion;
-					sx += 1.0;
+					sx += 1.0f;
 				}
 				while (sx < sx2);
 				sy += 1.0f;
@@ -3549,14 +3549,14 @@ BGD_DECLARE(void) gdImageCopyResampled (gdImagePtr dst,
 				blue /= alpha_sum;
 			}
 			/* Clamping to allow for rounding errors above */
-			if (red > 255.0) {
-				red = 255.0;
+			if (red > 255.0f) {
+				red = 255.0f;
 			}
-			if (green > 255.0) {
-				green = 255.0;
+			if (green > 255.0f) {
+				green = 255.0f;
 			}
 			if (blue > 255.0f) {
-				blue = 255.0;
+				blue = 255.0f;
 			}
 			if (alpha > gdAlphaMax) {
 				alpha = gdAlphaMax;
